added event and attendee policies

+++ app/Http/Controllers/Api/AttendeeController.php
@@ -7,6 +7,8 @@ use App\Http\Resources\AttendeeResource;
 use App\Http\Traits\CanLoadRelationships;
 use App\Models\Attendee;
 use App\Models\Event;
+use Illuminate\Auth\Access\HandlesAuthorization;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Http\Request;
 use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
 use Illuminate\Support\Facades\Gate;
@@ -14,7 +16,7 @@ use Illuminate\Support\Facades\Gate;
 class AttendeeController extends Controller
 {
 
-    use CanLoadRelationships;
+    use CanLoadRelationships, AuthorizesRequests;
 
     private array $relations = ['user'];
 
@@ -24,6 +26,8 @@ class AttendeeController extends Controller
 
     public function index(Event $event): AnonymousResourceCollection
     {
+        $this->authorize('viewAny', Attendee::class);
+
         $attendees = $this->loadRelationships($event->attendees()->latest());
 
         return AttendeeResource::collection($attendees->paginate());
@@ -34,6 +38,8 @@ class AttendeeController extends Controller
      */
     public function store(Request $request, Event $event): AttendeeResource
     {
+        $this->authorize('create', Attendee::class);
+
         $attendee = $this->loadRelationships($event->attendees()->create([
             'user_id' => $request->user()->id,
         ]));
@@ -46,6 +52,8 @@ class AttendeeController extends Controller
      */
     public function show(Event $event, Attendee $attendee): AttendeeResource
     {
+        $this->authorize('view', $attendee);
+
         return new AttendeeResource($this->loadRelationships($attendee));
     }
 
@@ -62,8 +70,10 @@ class AttendeeController extends Controller
      */
     public function destroy(Event $event, Attendee $attendee)
     {
+        $this->authorize('delete', $attendee);
+
         Gate::authorize('delete-attendee', [$event, $attendee]);
-        
+
         $attendee->delete();
 
         return response(status: 204);
